refactor: 미디어 폴더명 대소문자 변경 및 import에러 수정
