+++ src/components/Media/NavigationControls.tsx
